% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncoding_load.R
\name{load_noncoding_hotspots}
\alias{load_noncoding_hotspots}
\title{Load Noncoding Hotspot Database}
\usage{
load_noncoding_hotspots(verbose = TRUE)
}
\arguments{
\item{verbose}{verbose? (boolean)}
}
\value{
a tibble  containing describing noncoding hotspots
}
\description{
Loads a tibble  containing describing noncoding hotspots
}
\examples{
load_noncoding_hotspots()
}
